# use flat array for field indexing instead of hashmap