--- a/src/semantics.rs
+++ b/src/semantics.rs
@@ -9,54 +9,97 @@ pub trait Validate {
 }
 
 pub trait FieldCounter {
-    fn add_to_counter(&self, map: &mut HashMap<String, usize>);
-    fn remove_from_counter(&self, map: &mut HashMap<String, usize>);
+    fn add_to_counter(&mut self, fields: &mut Vec<(String, usize)>, map: &mut HashMap<String, usize>);
+    fn remove_from_counter(&mut self, fields: &mut Vec<(String, usize)>, map: &mut HashMap<String, usize>);
+    fn fix_lhs_index(&mut self, map: &HashMap<String, usize>);
 }
 
 impl FieldCounter for Expression {
-    fn add_to_counter(&self, map: &mut HashMap<String, usize>) {
-        match self {
-            Expression::Logical(l) => match l.as_ref() {
+    fn add_to_counter(&mut self, fields: &mut Vec<(String, usize)>, map: &mut HashMap<String, usize>) {
+            match self {
+            Expression::Logical(l) => match l.as_mut() {
                 LogicalExpression::And(l, r) => {
-                    l.add_to_counter(map);
-                    r.add_to_counter(map);
+                    l.add_to_counter(fields, map);
+                    r.add_to_counter(fields, map);
                 }
                 LogicalExpression::Or(l, r) => {
-                    l.add_to_counter(map);
-                    r.add_to_counter(map);
+                    l.add_to_counter(fields,map);
+                    r.add_to_counter(fields, map);
                 }
                 LogicalExpression::Not(r) => {
-                    r.add_to_counter(map);
+                    r.add_to_counter(fields, map);
                 }
             },
             Expression::Predicate(p) => {
-                *map.entry(p.lhs.var_name.clone()).or_default() += 1;
+                // 1. fields: increment counter for field
+                // 2. lhs: assign field index to the LHS
+                // 3. map: maintain the fields map: {field_name : field_index}
+                if let Some(index) = map.get(&p.lhs.var_name) {
+                    fields[*index].1 += 1;
+                    p.lhs.index = *index;
+                } else {
+                    fields.push((p.lhs.var_name.clone(), 1));
+                    map.insert(p.lhs.var_name.clone(), fields.len()-1);
+                    p.lhs.index = fields.len()-1;
+                }
             }
         }
     }
 
-    fn remove_from_counter(&self, map: &mut HashMap<String, usize>) {
+    fn remove_from_counter(&mut self, fields: &mut Vec<(String, usize)>, map: &mut HashMap<String, usize>) {
         match self {
-            Expression::Logical(l) => match l.as_ref() {
+            Expression::Logical(l) => match l.as_mut() {
                 LogicalExpression::And(l, r) => {
-                    l.remove_from_counter(map);
-                    r.remove_from_counter(map);
+                    l.remove_from_counter(fields, map);
+                    r.remove_from_counter(fields, map);
                 }
                 LogicalExpression::Or(l, r) => {
-                    l.remove_from_counter(map);
-                    r.remove_from_counter(map);
+                    l.remove_from_counter(fields, map);
+                    r.remove_from_counter(fields, map);
                 }
                 LogicalExpression::Not(r) => {
-                    r.remove_from_counter(map);
+                    r.remove_from_counter(fields, map);
                 }
             },
             Expression::Predicate(p) => {
-                let val = map.get_mut(&p.lhs.var_name).unwrap();
-                *val -= 1;
-
-                if *val == 0 {
+                // field must be in hashmap and fields array
+                let index: usize = *map.get(&p.lhs.var_name).unwrap();
+                // 1. decrement counter of field
+                // 2. for field removing, swap another field to fill the slot.
+                // 3. re-assign index in fields_map for swapped field in array.
+                fields[index].1 -= 1;
+                if fields[index].1 == 0 {
+                    fields.swap_remove(index);
                     assert!(map.remove(&p.lhs.var_name).is_some());
+                    // for field who been swapped to fill the empty slot in fields array,
+                    // we need to fix its index in fields_map
+                    if index < fields.len() {
+                        *map.get_mut(&fields[index].0).unwrap() = index;
+                    }
                 }
+            }
+        }
+    }
+
+    // this may run only after remove_from_counter, since existing field in field array
+    // may change its postion after that. run this to fix lhs's field index.
+    fn fix_lhs_index(&mut self, map: &HashMap<String, usize>) {
+        match self {
+            Expression::Logical(l) => match l.as_mut() {
+                LogicalExpression::And(l, r) => {
+                    l.fix_lhs_index(map);
+                    r.fix_lhs_index(map);
+                }
+                LogicalExpression::Or(l, r) => {
+                    l.fix_lhs_index(map);
+                    r.fix_lhs_index(map);
+                }
+                LogicalExpression::Not(r) => {
+                    r.fix_lhs_index(map);
+                }
+            },
+            Expression::Predicate(p) => {
+                p.lhs.index = *map.get(&p.lhs.var_name).unwrap();
             }
         }
     }