--- a/t/02-bugs.t
+++ b/t/02-bugs.t
@@ -39,9 +39,9 @@ __DATA__
                 "\xfc\x80\x80\x80\x80\xaf",
             }
 
-            local c = context.new(s)
+            local c = context.new(s, #r:get_fields())
             for _, v in ipairs(BAD_UTF8) do
-                local ok, err = c:add_value("http.path", v)
+                local ok, err = c:add_value(1, "http.path", v)
                 ngx.say(err)
             end
         }
@@ -71,8 +71,8 @@ invalid utf-8 sequence of 1 bytes from index 0
 
             s:add_field("http.path", "String")
 
-            local c = context.new(s)
-            assert(c:add_value("http.path", "\x00"))
+            local c = context.new(s, #r:get_fields())
+            assert(c:add_value(1, "http.path", "\x00"))
             ngx.say("ok")
         }
     }
@@ -146,8 +146,8 @@ ok
             assert(r:add_matcher(0, "a921a9aa-ec0e-4cf3-a6cc-1aa5583d150c",
                                  "http.body =^ \"world\""))
 
-            local c = context.new(s)
-            c:add_value("http.body", "hello\x00world")
+            local c = context.new(s, #r:get_fields())
+            c:add_value(1, "http.body", "hello\x00world")
 
             local matched = r:execute(c)
             ngx.say(matched)
@@ -156,7 +156,7 @@ ok
             ngx.say(uuid)
 
             c:reset()
-            c:add_value("http.body", "world\x00hello")
+            c:add_value(1, "http.body", "world\x00hello")
 
             local matched = r:execute(c)
             ngx.say(matched)