… Cargo.lock
… Cargo.toml
… benches/data.json
… t/01-sanity.t
… t/03-contains.t
… t/04-rawstr.t
… t/05-equals.t
… t/07-in_notin.t
… t/08-equals.t
… t/09-not.t