To resolve [KAG-5155](https://konghq.atlassian.net/browse/KAG-5155)