@ProBrian 2024-10-30
Local benchmark result attached (MacOS M3 chip with 36GB mem), comparing with PR branch and main branch
```
 Running benches/build.rs (target/release/deps/build-3f0a7ade35c0c7a4)
New: Build Router            time:   [3.0939 ms 3.1021 ms 3.1114 ms]
Old: Build Router            time:   [3.1140 ms 3.1375 ms 3.1765 ms]
                        change: [+1.2814% +2.1473% +3.3212%] (p = 0.00 < 0.05)

 
     Running benches/match_mix.rs (target/release/deps/match_mix-8ebfb4b8ad1a6fd2)
New: Match                   time:   [36.678 ns 36.807 ns 36.991 ns]
Old: Match                   time:   [56.656 ns 56.897 ns 57.178 ns]
                        change: [+10.838% +11.498% +12.176%] (p = 0.00 < 0.05)

 
     Running benches/misc_match.rs (target/release/deps/misc_match-ef04ce8d11de66a9)
  New: route match all         time:   [103.81 ns 104.00 ns 104.17 ns]
  Old:  route match all         time:   [89.362 ns 89.910 ns 90.633 ns]

 
New: route mismatch all      time:   [123.28 ns 123.52 ns 123.80 ns]
Old: route mismatch all      time:   [225.65 ns 226.23 ns 226.94 ns]

 
New: route matchers batch create and delete
                        time:   [20.357 ms 20.448 ms 20.555 ms]
Old: route matchers batch create and delete
                        time:   [19.954 ms 20.100 ms 20.272 ms]

 
     Running benches/not_match_mix.rs (target/release/deps/not_match_mix-8894a45acea74fc2)
New: Doesn't Match           time:   [13.551 ms 13.624 ms 13.715 ms]
Old: Doesn't Match           time:   [21.577 ms 22.097 ms 22.635 ms]

 
     Running benches/string.rs (target/release/deps/string-4f7ea951a6e3683d)
New: Doesn't Match           time:   [10.536 ms 10.571 ms 10.615 ms]
Old: Doesn't Match           time:   [14.018 ms 14.069 ms 14.125 ms]

 
     Running benches/test.rs (target/release/deps/test-1f097a0352e64a6c)
New: Doesn't Match           time:   [35.154 ms 35.342 ms 35.552 ms]
Old: Doesn't Match           time:   [35.460 ms 35.740 ms 36.037 ms]

```