@ProBrian 2024-12-23
> Please resolve conflicts

Conflicts caused due to the revert introduced by https://github.com/Kong/atc-router/pull/284, we should figure out if that reverted change will be committed back after the investigation of performance.